Develop

+++ dynamic-mapping-ui/src/extension/extension-card/extension-card.component.ts
@@ -22,8 +22,9 @@ import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { IManagedObject } from '@c8y/client';
 import { AlertService } from '@c8y/ngx-components';
-import { ExtensionStatus } from '../../shared';
+import { ConfirmationModalComponent, ExtensionStatus } from '../../shared';
 import { ExtensionService } from '../share/extension.service';
+import { BsModalRef, BsModalService } from 'ngx-bootstrap/modal';
 
 @Component({
   selector: 'd11r-mapping-extension-card',
@@ -31,7 +32,6 @@ import { ExtensionService } from '../share/extension.service';
 })
 export class ExtensionCardComponent implements OnInit {
   @Input() app: IManagedObject;
-  @Input() loaded: any = true;
   @Output() appDeleted: EventEmitter<void> = new EventEmitter();
   ExtensionStatus = ExtensionStatus;
   external: boolean = true;
@@ -40,15 +40,16 @@ export class ExtensionCardComponent implements OnInit {
     private extensionService: ExtensionService,
     private alertService: AlertService,
     private router: Router,
-    private activatedRoute: ActivatedRoute
+    private activatedRoute: ActivatedRoute,
+    public bsModalService: BsModalService
   ) {}
 
   async ngOnInit() {
-    this.external = this.app?.external;
+    this.external = this.app?.['external'];
   }
 
   async detail() {
-    // this.router.navigateByUrl(`/sag-ps-pkg-dynamic-mapping/extensions/${this.app.id}`);
+    // this.router.navigateByUrl(`/sag-ps-pkg-dynamic-mapping/extension/${this.app.id}`);
     this.router.navigate(['properties/', this.app.id], {
       relativeTo: this.activatedRoute
     });
@@ -56,13 +57,36 @@ export class ExtensionCardComponent implements OnInit {
   }
 
   async delete() {
-    try {
-      await this.extensionService.deleteExtension(this.app);
-      this.appDeleted.emit();
-    } catch (ex) {
-      if (ex) {
-        this.alertService.addServerFailure(ex);
-      }
-    }
+    const initialState = {
+        title: 'Delete mapping extension',
+        message:
+          'You are about to delete a mapping extension. Do you want to proceed?',
+        labels: {
+          ok: 'Delete',
+          cancel: 'Cancel'
+        }
+      };
+      const confirmDeletionModalRef: BsModalRef = this.bsModalService.show(
+        ConfirmationModalComponent,
+        { initialState }
+      );
+      confirmDeletionModalRef.content.closeSubject.subscribe(
+        async (confirmation: boolean) => {
+          console.log('Confirmation result:', confirmation);
+          if (confirmation) {
+            try {
+                await this.extensionService.deleteExtension(this.app);
+                this.appDeleted.emit();
+              } catch (ex) {
+                if (ex) {
+                  this.alertService.addServerFailure(ex);
+                }
+              }
+          }
+          confirmDeletionModalRef.hide();
+        }
+      );
+
+
   }
 }
